backend/requirements.txt
# Learno Backend Requirements

# Web framework
fastapi==0.109.0
uvicorn==0.27.0

# HTTP client
httpx==0.26.0

# OpenAI
openai==1.12.0

# Environment
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
email-validator>=2.1.0

# Image processing
pillow==10.2.0

# Database
sqlalchemy>=2.0.0

# Auth
bcrypt>=4.1.0
pyjwt>=2.8.0
python-multipart>=0.0.6

# Rate limiting
slowapi>=0.1.9
